Change edge type to index based as well
Much easier to handle undirected edges and now we don't have tu use
hashmaps anymore.

Additionally removed edge_attrs and node_attrs in favour of just always
using attrs. This is in one way not great, as the code will be a bit
messer when we only need one. Maybe should reverse this. But in one way
it is good to be able to abstract away the underlying attrs structs and
have them be the same for edges and nodes.